requirement.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.0.7
django-extensions==3.2.0
django-filter==22.1
django-sequences==2.7
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
mssql-django==1.1.3
mysqlclient==2.1.1
Naked==0.1.31
Pillow==9.2.0
pyodbc==4.0.34
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
status==0.2.5
tzdata==2022.2
urllib3==1.26.12
zipp==3.8.1
